preflight: Factor tray download/extraction code
The code to download/extract the tray on Windows and macOS is
duplicated. This commit moves it to a single helper.
